refactor: refactor tests
