+++ src/sumOfOther.js
